+++ __tests__/middleware/collectionRedirects.test.tsx
@@ -13,11 +13,11 @@ jest.mock("next/server", () => {
 
 let request: NextRequest;
 
-it("redirects collection_keywords to q", async () => {
+it("redirects collection_keywords to q", () => {
   const request = {
     nextUrl: new URL("http://localhost/collections?collection_keywords=test"),
   } as NextRequest;
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.redirect).toHaveBeenCalledWith(
     "http://localhost/collections?q=test",
@@ -26,13 +26,13 @@ it("redirects collection_keywords to q", async () => {
   expect(response).toBe("redirect response");
 });
 
-it("transforms collection_keywords to q and maintains other parameters", async () => {
+it("transforms collection_keywords to q and maintains other parameters", () => {
   const request = {
     nextUrl: new URL(
       "http://localhost/collections?collection_keywords=test&sort=title-asc&page=23"
     ),
   } as NextRequest;
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.redirect).toHaveBeenCalledWith(
     "http://localhost/collections?sort=title-asc&page=23&q=test",
@@ -41,23 +41,23 @@ it("transforms collection_keywords to q and maintains other parameters", async (
   expect(response).toBe("redirect response");
 });
 
-it("goes through unchanged if no collection_keywords= param", async () => {
+it("goes through unchanged if no collection_keywords= param", () => {
   request = {
     nextUrl: new URL("http://localhost/collections?sort=title-asc"),
   } as NextRequest;
 
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.next).toHaveBeenCalled();
   expect(response).toBe("next response");
 });
 
-it("redirects existing collection slug to uuid", async () => {
+it("redirects existing collection slug to uuid", () => {
   request = {
     nextUrl: new URL("http://localhost/collections/script-b"),
   } as NextRequest;
 
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.redirect).toHaveBeenCalledWith(
     "http://localhost/collections/5af6c940-c45b-013c-85aa-0242ac110005",
@@ -66,12 +66,12 @@ it("redirects existing collection slug to uuid", async () => {
   expect(response).toBe("redirect response");
 });
 
-it("redirects non-existent collection slug to title search", async () => {
+it("redirects non-existent collection slug to title search", () => {
   request = {
     nextUrl: new URL("http://localhost/collections/hello-im-not-a-real-slug"),
   } as NextRequest;
 
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.redirect).toHaveBeenCalledWith(
     "http://localhost/search/index?q=Hello+Im+Not+A+Real+Slug",
@@ -80,14 +80,14 @@ it("redirects non-existent collection slug to title search", async () => {
   expect(response).toBe("redirect response");
 });
 
-it("redirects a deprecated division slug to correct slug", async () => {
+it("redirects a deprecated division slug to correct slug", () => {
   request = {
     nextUrl: new URL(
       "http://localhost/divisions/schomburg-center-for-research-in-black-culture-jean-blackwell-hutson-research-and-reference-division"
     ),
   } as NextRequest;
 
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.redirect).toHaveBeenCalledWith(
     "http://localhost/divisions/schomburg-center-for-research-in-black-culture-jean-blackwell-hutson-research",
@@ -96,14 +96,14 @@ it("redirects a deprecated division slug to correct slug", async () => {
   expect(response).toBe("redirect response");
 });
 
-it("redirects another deprecated division slug to correct slug", async () => {
+it("redirects another deprecated division slug to correct slug", () => {
   request = {
     nextUrl: new URL(
       "http://localhost/divisions/the-miriam-and-ira-d-wallach-division-of-art-prints-and-photographs-picture-collection"
     ),
   } as NextRequest;
 
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.redirect).toHaveBeenCalledWith(
     "http://localhost/divisions/the-miriam-and-ira-d-wallach-division-of-art-prints-and-photographs-picture",
@@ -112,25 +112,25 @@ it("redirects another deprecated division slug to correct slug", async () => {
   expect(response).toBe("redirect response");
 });
 
-it("does not redirect a current division slug", async () => {
+it("does not redirect a current division slug", () => {
   request = {
     nextUrl: new URL("http://localhost/divisions/billy-rose-theatre-division"),
   } as NextRequest;
 
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.next).toHaveBeenCalled();
   expect(response).toBe("next response");
 });
 
-it("passes a collection uuid through", async () => {
+it("passes a collection uuid through", () => {
   request = {
     nextUrl: new URL(
       "http://localhost/collections/bc920670-c5f9-012f-63ad-58d385a7bc34"
     ),
   } as NextRequest;
 
-  const response = await middleware(request);
+  const response = middleware(request);
 
   expect(NextResponse.next).toHaveBeenCalled();
   expect(response).toBe("next response");