+++ app/src/components/pages/collectionPage/collectionPage.tsx
@@ -37,6 +37,7 @@ import CollectionMetadata, {
 import CollectionSearchParamsType from "@/src/types/CollectionSearchParams";
 import { SearchResultsType } from "@/src/types/SearchResultsType";
 import { CollectionModel } from "@/src/models/collection";
+import { useSubcollectionRedirect } from "@/src/hooks/useSubcollectionRedirect";
 
 type CollectionPageProps = {
   searchResults: SearchResultsType;
@@ -76,6 +77,8 @@ const CollectionPage = ({
     push(`${pathname}?${queryString}`, { scroll: false });
   };
   const [isLoaded, setIsLoaded] = useState(false);
+
+  useSubcollectionRedirect();
 
   useEffect(() => {
     setIsLoaded(true);
@@ -150,7 +153,10 @@ const CollectionPage = ({
           paddingRight: { base: "m", xl: "s" },
         }}
       >
-        <ActiveFilters searchManager={collectionSearchManager} />
+        <ActiveFilters
+          searchManager={collectionSearchManager}
+          allFilters={searchResults.availableFilters}
+        />
         <CollectionMetadata data={new CollectionModel(collectionData)} />
         <Flex
           gap="xxl"