+++ app/src/components/pages/searchPage/searchPage.tsx
@@ -22,6 +22,7 @@ import NoResultsFound from "../../results/noResultsFound";
 import SearchCardGridLoading from "../../grids/searchCardGridLoading";
 import BackToTopLink from "../../backToTopLink/backToTopLink";
 import { SearchResultsType } from "@/src/types/SearchResultsType";
+import { useSubcollectionRedirect } from "@/src/hooks/useSubcollectionRedirect";
 
 const SearchPage = ({
   searchResults,
@@ -76,6 +77,8 @@ const SearchPage = ({
     isFirstLoad.current = true;
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [searchResults]);
+
+  useSubcollectionRedirect();
 
   return (
     <Box id="mainContent">
@@ -157,7 +160,10 @@ const SearchPage = ({
           marginTop: searchResults.results?.length > 0 ? "0" : "m",
         }}
       >
-        <ActiveFilters searchManager={searchManager} />
+        <ActiveFilters
+          searchManager={searchManager}
+          allFilters={searchResults.availableFilters}
+        />
         <Flex
           sx={{
             [`@media screen and (min-width: ${headerBreakpoints.lgMobile}px)`]: