+++ app/src/utils/utils.ts
@@ -176,9 +176,6 @@ export const filterStringToCollectionApiFilterString = (filters: string) => {
         if (!isValidFilter(name)) {
           return null;
         }
-        if (name === "subcollection" || name === "collection") {
-          return `${name}=${value}`;
-        }
         return `${encodeURIComponent(name)}=${encodeURIComponent(value)}`;
       })
       .filter(Boolean);
@@ -264,6 +261,12 @@ export const getTitleWithHighlights = (highlights, title) => {
   return titleHighlight ? replaceEmWithMark(titleHighlight.text) : title;
 };
 */
+
+// DC uuids are not uuid v1-5 compliant– they're v0 (kind of?), which this regex tests
+const dcUuidRegex = /^[\da-f]{8}-([\da-f]{4}-){3}[\da-f]{12}$/i;
+export const isDCUuid = (identifier) => {
+  return dcUuidRegex.test(identifier);
+};
 
 export const deSlugify = (slug: string): string => {
   return slug.replace(/-/g, " ").replace(/\b\w/g, (c) => c.toUpperCase());