+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Connected collection landing page (metadata) to API (DR-3362)
 - Added collection structure component (DR-3545)
 - Added deprecated division slug redirects
+- Added clientside subcollection redirects (DR-3616)
+
+### Updated
+
+- Updated collection and subcollection filter value format (DR-3515)
 
 ## [0.4.1] 2025-05-16